Move config objects into subfolder
